Render even if the frame size changes
Fixes #23 (almost, there are still very minor glitches when the
requested frame count goes up, but barely noticeable due to only being 1
or 2 samples).

Note that the large frame counts on iOS (e.g. 940 compared to 512 on macOS)
also cause parameter changes to take a bit longer to be audible. This is
probably an inherent issue since we need to snapshot the synthesizer graph
to avoid concurrency issues between the audio thread and the main thread
which controls the UI, just something to be aware of.